+++ rxros_tf/CMakeLists.txt
@@ -1,17 +1,17 @@
 cmake_minimum_required(VERSION 2.8.3)
-project(rxros)
+project(rxros_tf)
 
 find_package(catkin REQUIRED COMPONENTS
-  roscpp
-  rxcpp_vendor
+  rxros
+  tf
 )
 
 catkin_package(
   INCLUDE_DIRS
     include
   CATKIN_DEPENDS
-    roscpp
-    rxcpp_vendor
+    rxros
+    tf
 )
