+++ rxros_tf/package.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>rxros</name>
+  <name>rxros_tf</name>
   <version>0.1.0</version>
-  <description>The RxROS package</description>
+  <description>Extensions to RxROS for working with TF.</description>
 
   <author>Henrik Larsen</author>
   <maintainer email="wasowski@itu.dk">Andrzej Wasowski</maintainer>
@@ -13,9 +13,9 @@
 
   <buildtool_depend>catkin</buildtool_depend>
 
-  <depend>roscpp</depend>
+  <build_depend>rxros</build_depend>
 
-  <build_depend>rxcpp_vendor</build_depend>
+  <build_export_depend>rxros</build_export_depend>
 
-  <build_export_depend>rxcpp_vendor</build_export_depend>
+  <depend>tf</depend>
 </package>
